# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample HTTP package with Swagger annotations.",
        "title": "VK-Filmoteka",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/http/v1/actor": {
            "get": {
                "description": "Можно задавать разные query",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Получение актера",
                "parameters": [
                    {
                        "type": "string",
                        "description": "текущая страница",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "кол-во актеров на странице",
                        "name": "page-count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Вернуть вместе со связями (film)",
                        "name": "connection",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "получаемые актеры",
                        "schema": {
                            "$ref": "#/definitions/appDto.ActorGetByQueryResult"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Доступно только админам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Обновление актера [Админы]",
                "parameters": [
                    {
                        "description": "Данные актера",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Actor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные актера",
                        "schema": {
                            "$ref": "#/definitions/model.Actor"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "post": {
                "description": "Доступно только админам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Создание актера [Админы]",
                "parameters": [
                    {
                        "description": "Данные актера",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appDto.CreateActorUseCaseDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные созданного актера",
                        "schema": {
                            "$ref": "#/definitions/model.Actor"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Доступно только админам, ничего ответом не возвращает",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Удаление актера [Админы]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id удаляемого пользователья",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        },
        "/http/v1/actor/add-film": {
            "post": {
                "description": "Доступно только админам",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "actor"
                ],
                "summary": "Обновление актера [Админы]",
                "parameters": [
                    {
                        "description": "Данные id для связывания актера и фильма",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.AddFilmToActorDto"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        },
        "/http/v1/auth/login": {
            "post": {
                "description": "Ответом при успешном Логине получаем свои данные",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Логин пользователя",
                "parameters": [
                    {
                        "description": "Данные пользователя",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appDto.LoginUseCaseDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные пользователя",
                        "schema": {
                            "$ref": "#/definitions/appDto.ResponseUserDto"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        },
...
        "/http/v1/auth/refresh": {
            "post": {
                "description": "Ответом при успешном Логине получаем свои данные",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновление access токена пользователя",
                "responses": {
                    "200": {
                        "description": "Данные созданного пользователя",
                        "schema": {
                            "$ref": "#/definitions/appDto.ResponseUserDto"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        },
        "/http/v1/auth/registration": {
            "post": {
                "description": "Ответом при успешном регистрация получаем свои данные",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация пользователя",
                "parameters": [
                    {
                        "description": "Данные нового пользователя",
                        "name": "reg",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/appDto.RegistrationUseCaseDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Данные созданного пользователя",
                        "schema": {
                            "$ref": "#/definitions/appDto.ResponseUserDto"
                        }
                    },
                    "404": {
                        "description": "Ошибка 404",
                        "schema": {
                            "$ref": "#/definitions/appErrors.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "aggregate.ActorAggregate": {
            "type": "object",
            "properties": {
                "actor": {
                    "$ref": "#/definitions/model.Actor"
                },
                "films": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Film"
                    }
                }
            }
        },
...
        "appDto.ActorGetByQueryResult": {
            "type": "object",
            "properties": {
                "actors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/aggregate.ActorAggregate"
                    }
                },
                "pageCount": {
                    "type": "integer"
                }
            }
        },
        "appDto.CreateActorUseCaseDto": {
            "type": "object",
            "required": [
                "birthday",
                "gender",
                "name"
            ],
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
...
        "appDto.LoginUseCaseDto": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 35,
                    "minLength": 8
                }
            }
        },
        "appDto.RegistrationUseCaseDto": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 3
                },
                "password": {
                    "type": "string",
                    "maxLength": 35,
                    "minLength": 8
                }
            }
        },
        "appDto.ResponseUserDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "appErrors.ResponseError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "dto.AddFilmToActorDto": {
            "type": "object",
            "required": [
                "actorId",
                "filmIds"
            ],
            "properties": {
                "actorId": {
                    "type": "string"
                },
                "filmIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.Actor": {
            "type": "object",
            "required": [
                "birhday",
                "gender",
                "id",
                "name"
            ],
            "properties": {
                "birhday": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                }
            }
        },
        "model.Film": {
            "type": "object",
            "required": [
                "id",
                "name",
                "release"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 1000
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 1
                },
                "rate": {
                    "type": "number",
                    "maximum": 10,
                    "minimum": 0
                },
                "release": {
                    "type": "string"
                }
            }
        }
    }
}